+++ tests/ops/integration_tests/saas/test_adobe_sign_task.py
@@ -1,11 +1,11 @@
 import pytest
 
-from fides.api.models.policy import Policy
 from tests.ops.integration_tests.saas.connector_runner import ConnectorRunner
 
 
+@pytest.mark.skip(reason="Currently unable to test OAuth2 connectors")
 @pytest.mark.integration_saas
-class TestAdobe_SignConnector:
+class TestAdobeSignConnector:
     def test_connection(self, adobe_sign_runner: ConnectorRunner):
         adobe_sign_runner.test_connection()
 
@@ -16,5 +16,5 @@ class TestAdobe_SignConnector:
             access_policy=policy, identities={"email": adobe_sign_identity_email}
         )
 
-        # for users in access_results["adobe_sign_instance:users"]:
-        #     assert users["email"] == adobe_sign_identity_email
+        for users in access_results["adobe_sign_instance:users"]:
+            assert users["email"] == adobe_sign_identity_email