+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The types of changes are:
 
 ## [Unreleased](https://github.com/ethyca/fides/compare/2.17.0...main)
 
+### Added
+- Access support for Adobe Sign [#3504](https://github.com/ethyca/fides/pull/3504)
+
 ### Changed
 - Simplified the file structure for HTML DSR packages [#3848](https://github.com/ethyca/fides/pull/3848)
 
@@ -27,7 +30,6 @@ The types of changes are:
 - Tab component for `fides-js` [#3782](https://github.com/ethyca/fides/pull/3782)
 - Added toast for successfully linking an existing integration to a system [#3826](https://github.com/ethyca/fides/pull/3826)
 - Various other UI components for `fides-js` to support upcoming TCF modal [#3803](https://github.com/ethyca/fides/pull/3803)
-- Prefetches API calls as part of Fides.js [#3698](https://github.com/ethyca/fides/pull/3698)
 - Allow items in taxonomy to be enabled or disabled [#3844](https://github.com/ethyca/fides/pull/3844)
 
 ### Developer Experience
@@ -43,7 +45,6 @@ The types of changes are:
 - Sort system cards alphabetically by name on "View systems" page [#3781](https://github.com/ethyca/fides/pull/3781)
 - Update admin ui to use new integration delete route [#3785](https://github.com/ethyca/fides/pull/3785)
 - Pinned `pymssql` and `cython` dependencies to avoid build issues on ARM machines [#3829](https://github.com/ethyca/fides/pull/3829)
-- Disable server-side geolocation for Fides.js [#3850](https://github.com/ethyca/fides/pull/3850)
 
 ### Removed